requirements.txt

pypdf
python-docx
python-pptx
markdown
beautifulsoup4
pillow
pytesseract
faster-whisper
youtube-transcript-api
pytube
numpy
scikit-learn
rank-bm25
sentence-transformers
google-generativeai
streamlit
python-dotenv


